requirements.txt
biopython
pyfastx
pandas